requirements.txt
python-telegram-bot==20.0
feedparser==6.0.10
requests==2.31.0
openai==1.3.0
apscheduler==3.10.4
pyyaml==6.0.1
python-dotenv==1.0.0
